requirements.txt
djangorestframework
Django
djangorestframework-simplejwt